TG-23: close form after send request. Loader for index
