--- src/Components/SuperAdmins/FormAdd/AddSAdmin.jsx
@@ -1,112 +1,0 @@
-import React, { useState } from 'react';
-import styles from './addSAdmin.module.css';
-
-const AddSuperAdmin = ({ show, closeForm, setShowModal, setShowTitle, addItem }) => {
-  if (!show) {
-    return null;
-  }
-
-  const [userInput, setUserInput] = useState({
-    firstName: '',
-    lastName: '',
-    email: '',
-    password: '',
-    active: ''
-  });
-
-  const onChange = (e) => {
-    setUserInput({ ...userInput, [e.target.name]: e.target.value });
-  };
-
-  const onSubmit = async (e) => {
-    e.preventDefault();
-
-    setUserInput({
-      firstName: '',
-      lastName: '',
-      email: '',
-      password: '',
-      active: ''
-    });
-
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-type': 'application/json'
-      },
-      body: JSON.stringify({
-        firstName: userInput.firstName,
-        lastName: userInput.lastName,
-        email: userInput.email,
-        password: userInput.password,
-        active: userInput.active
-      })
-    };
-
-    try {
-      const response = await fetch(`${process.env.REACT_APP_API_URL}/super-admins`, options);
-      const data = await response.json();
-      if (response.status !== 200 && response.status !== 201) {
-        setShowModal(true);
-        setShowTitle(data.message);
-      }
-      setShowTitle('Super Admin Created');
-      setShowModal(true);
-      addItem(data.data);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return (
-    <div className={styles.container}>
-      <form onSubmit={onSubmit}>
-        <h2>Add SuperAdmin</h2>
-        <div>
-          <label>Name</label>
-          <input
-            type="text"
-            name="firstName"
-            value={userInput.firstName}
-            onChange={onChange}
-          ></input>
-        </div>
-        <div>
-          <label>Last Name</label>
-          <input type="text" name="lastName" value={userInput.lastName} onChange={onChange}></input>
-        </div>
-        <div>
-          <label>Email</label>
-          <input type="text" name="email" value={userInput.email} onChange={onChange}></input>
-        </div>
-        <div>
-          <label>Password</label>
-          <input
-            type="password"
-            name="password"
-            value={userInput.password}
-            onChange={onChange}
-          ></input>
-        </div>
-        <div>
-          <label>Active</label>
-          <input type="text" name="active" value={userInput.active} onChange={onChange}></input>
-        </div>
-        <div>
-          <input
-            type="submit"
-            value="Confirm"
-            onSubmit={() => {
-              setShowModal(true);
-            }}
-          ></input>
-        </div>
-        <div>
-          <button onClick={closeForm}>Close</button>
-        </div>
-      </form>
-    </div>
-  );
-};
-
-export default AddSuperAdmin;
