+++ src/Components/SuperAdmins/FormAdd/addSAdmin.module.css
@@ -29,7 +29,7 @@
   align-items: center;
   width: 35%;
   height: 80%;
-  background-color: #fff;
+  background-color: rgb(255, 255, 255);
   margin-left: 5%;
   border-radius: 10px;
   gap: 15px
@@ -38,36 +38,36 @@
 .container input[type="submit"]{
   padding: 10px 15px;
   margin-right: 10px;
-  background-color: #AACE9B;
-  border: black solid 1px;
+  background-color: rgb(170, 206, 155);
+  border: rgb(0, 0, 0) solid 1px;
   border-radius: 5px;
-  color: #373867;
+  color: rgb(55, 56, 103);
   box-shadow: 3px 3px 10px
 }
 
 .container input[type="submit"]:hover{
-  background-color: #37781c;
-  color: #fff;
+  background-color: rgb(55, 120, 28);
+  color: rgb(255, 255, 255);
   transition-duration: 0,4s;
 }
 .container input[type="submit"]:active{
-  background-color: #265213;
+  background-color: rgb(38, 82, 19);
   transform: translateY(4px);
 }
 .container button{
   padding: 10px 15px;
-  background-color: #AACE9B;
-  border: black solid 1px;
+  background-color: rgb(170, 206, 155);
+  border: rgb(0, 0, 0) solid 1px;
   border-radius: 5px;
-  color: #373867;
+  color: rgb(55, 56, 103);
   box-shadow: 3px 3px 10px
 }
 .container button:hover{
-  background-color: #37781c;
-  color: #fff;
+  background-color: rgb(55, 120, 28);
+  color: rgb(255, 255, 255);
   transition-duration: 0,4s;
 }
 .container button:active{
-  background-color: #265213;
+  background-color: rgb(38, 82, 19);
   transform: translateY(4px);
 }
