+++ src/Components/SuperAdmins/ListItemSuperAdmin/listItemSAdmin.module.css
@@ -1,3 +1,7 @@
 .container td{
     padding-left: 50px;
 }
+
+tr:hover{
+    background-color: rgba(170, 206, 155, 0.705);
+}
