+++ src/Components/SuperAdmins/super-admins.module.css
@@ -25,3 +25,29 @@
   background-color: rgb(38, 82, 19);
   transform: translateY(4px);
 }
+
+.containerLoading {
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+
+.loader {
+  border: 10px solid #f3f3f3;
+  border-top: 10px solid #aace9b;
+  border-radius: 50%;
+  width: 80px;
+  height: 80px;
+  animation: spin 0.7s linear infinite;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
